Move PatternScan to its own class
